Feat(isAdm): função para detectar com mais praticidade se é ou não admin
