+++ src/bot/commands/compet/create.ts
@@ -1,16 +1,15 @@
 import { modal } from "@/bot/modals/compet/createMemberModal/createMemberModal";
 import { Command } from "../../structures/Command";
-import { makeNotAdminEmbed } from "@/bot/utils/embed/makeNotAdminEmbed";
+import { checkIfNotAdmin } from "@/bot/utils/embed/checkIfNotAdmin"
 
 export default new Command({
   name: "create",
   description: "Esse comando adiciona um novo competiano ao compet",
   run: async ({ interaction }) => {
-    const member = await interaction.guild?.members.fetch(interaction.user.id);
-    const isADM = member?.permissions.has("Administrator");
 
-    if (!isADM)
-      return makeNotAdminEmbed(interaction)
+    const isNotAdmin = await checkIfNotAdmin(interaction)
+    if ((isNotAdmin).isRight())
+      return isNotAdmin.value.response
 
     await interaction.showModal(modal);
   },
