+++ src/bot/errors/notAdminError.ts
@@ -1,0 +1,2 @@
+export class IsAdminError extends Error {
+}
